requirements.txt
streamlit
google-generativeai
PyMuPDF
requests